feat: add support for react-native v0.73 react-native v0.73 changed a few internal classes regarding ShadowNodes. I've updated the corresponding headers and function calls to the best of my knowledge.
The current state compiles, but I've commented out the "adopt" function call in `MGContentContainerComponentDescriptor.h`, because I couldn't get the types to work properly. It would be nice if someone could help me with that :)

Other than that, I want to thank the entire margelo team for creating this amazing library.
Please don't be discouraged to work this library. You've truly created something revolutionary, which has a bright future for the react-native space.